% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{fraq-package}
\alias{fraq-package}
\alias{fraq}
\title{fraq: fastq processing with TBB flow graphs}
\description{
fraq is a high-throughput toolkit for FASTQ processing. It uses a TBB flow
graph to coordinate concurrent I/O and compute, provides prebuilt kernels
for common tasks, and lets you extend it by writing custom kernels.
}
\details{
See individual function help pages for details on each kernel. The README
contains a short walkthrough with synthetic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/traversc/fraq}
  \item Report bugs at \url{https://github.com/traversc/fraq/issues}
}

}
\author{
\strong{Maintainer}: Travers Ching \email{traversc@gmail.com} (\href{https://orcid.org/0000-0002-5577-3516}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Yann Collet (Author of the bundled zstd library) [contributor, copyright holder]
  \item Facebook, Inc. (Copyright holder of the bundled zstd code) [copyright holder]
  \item Reichardt Tino (Contributor/copyright holder of bundled zstd code) [contributor, copyright holder]
  \item Skibinski Przemyslaw (Contributor/copyright holder of bundled zstd code) [contributor, copyright holder]
  \item Mori Yuta (Contributor/copyright holder of bundled zstd code) [contributor, copyright holder]
}

}
\keyword{internal}
