% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_documentation.R
\name{fraq_options}
\alias{fraq_options}
\title{Get or set FRAQ options}
\usage{
fraq_options(option, value = NULL)
}
\arguments{
\item{option}{Character string name of the option. Valid options are:
"blocksize", "fraq_compress_level", "zstd_compress_level",
"gzip_compress_level".}

\item{value}{Optional value to set the option to; if NULL, the current
value is returned.}
}
\value{
The current option value (if input \code{value} is \code{NULL}) or
previous option value.
}
\description{
Get or set FRAQ options.
}
\examples{
# Get current blocksize
fraq_options("blocksize")
# # Set blocksize to 16384
fraq_options("blocksize", 16384)

}
