% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_documentation.R
\name{fraq_trim_adapters}
\alias{fraq_trim_adapters}
\title{Trim adapters from FASTQ file(s)}
\usage{
fraq_trim_adapters(
  input,
  output,
  adapters,
  max_distance = 1L,
  filter_untrimmed = TRUE,
  nthreads = 1L
)
}
\arguments{
\item{input}{Character vector of one or more input FASTQ file paths.
Vectors must be the same length as \code{output} (e.g., R1 and R2 pairs).}

\item{output}{Character vector of output FASTQ file paths, same length as
\code{input}.}

\item{adapters}{Character vector of adapter sequences to trim. Adapters are
given priority based on the order they appear.}

\item{max_distance}{Integer maximum number of mismatches for adapter matching.}

\item{filter_untrimmed}{Logical; if TRUE, drop reads with no trim.}

\item{nthreads}{Integer number of threads.}
}
\value{
A data frame of counts of trimmed adapters.
}
\description{
Trim occurrences of adapter sequence(s) at the start of the
first fastq input. \code{input} and \code{output} must be vectors of the same
length (e.g., R1/R2 pairs).
Adapters will be trimmed only for the first fastq, but all inputs will be
filtered if \code{filter_untrimmed} is \code{TRUE}.
}
\examples{
r1 <- tempfile(fileext = ".fastq")
r2 <- tempfile(fileext = ".fastq")
generate_random_fastq(r1, n_reads = 1000, read_length = 100, 
    name_prefix = "read_R1_")
generate_random_fastq(r2, n_reads = 1000, read_length = 100, 
    name_prefix = "read_R2_")
out <- c(tempfile(fileext = ".fastq"), tempfile(fileext = ".fastq"))
adapters <- c("ACGT", "TGCA", "GTAC")
fraq_trim_adapters(c(r1, r2), out, adapters, max_distance = 1L, 
    filter_untrimmed = TRUE, nthreads = 1L)
}
